Refactor: SEC_03 dto 폴더 구조 수정 dto 하위에 폴더들 생성 후 분류 더 추가 예정
