+++ securities/src/main/java/com/baas/securities/dto/email/EmailValidationVerifyDTO.java
@@ -1,4 +1,4 @@
-package com.baas.securities.dto;
+package com.baas.securities.dto.email;
 
 
 import lombok.Getter;
