+++ securities/src/main/java/com/baas/securities/dto/stock/RealtimeStockDTO.java
@@ -1,4 +1,4 @@
-package com.baas.securities.dto;
+package com.baas.securities.dto.stock;
 
 import lombok.*;
