+++ securities/src/main/java/com/baas/securities/dto/stock/StockOrderDTO.java
@@ -1,9 +1,7 @@
-package com.baas.securities.dto;
+package com.baas.securities.dto.stock;
 
 import lombok.Getter;
 import lombok.Setter;
-
-import java.util.UUID;
 
 @Getter
 @Setter
