+++ securities/src/main/java/com/baas/securities/repository/dao/UserAgreementDao.java
@@ -1,0 +1,8 @@
+package com.baas.securities.repository.dao;
+
+import com.baas.securities.repository.UserAgreementRepository;
+import org.apache.ibatis.annotations.Mapper;
+
+@Mapper
+public interface UserAgreementDao extends UserAgreementRepository {
+}
